+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 #Change Log
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
+
+## [Unreleased]
+### Added
+- check-lv-usage.rb: added
+
 ## 0.0.1 - 2016-04-12
 ### Added
 - Initial release